refactor: split api.ts into focused auth.ts and proxy.ts modules
Move auth rate limiting + handleAuth to auth.ts and handleLocation +
handleWeather to proxy.ts. Each handler now receives corsHeaders as a
parameter instead of computing them internally.
