+++ api/src/index.ts
@@ -2,16 +2,9 @@ export { ChatRoom } from "./chat-room";
 export type * from "./types";
 
 import type { Env } from "./types";
-import {
-  handleCors,
-  handleWebSocket,
-  handleAuth,
-  handleConfig,
-  handleHealthCheck,
-  handleLocation,
-  handleWeather,
-  corsHeaders,
-} from "./api";
+import { handleCors, handleWebSocket, handleConfig, handleHealthCheck, corsHeaders } from "./api";
+import { handleAuth } from "./auth";
+import { handleLocation, handleWeather } from "./proxy";
 
 export default {
   async fetch(request: Request, env: Env, _ctx: ExecutionContext): Promise<Response> {
@@ -22,13 +15,13 @@ export default {
     } else if (url.pathname === "/ws") {
       return handleWebSocket(env, request);
     } else if (url.pathname === "/auth" && request.method === "POST") {
-      return handleAuth(env, request);
+      return handleAuth(env, request, corsHeaders(env, request));
     } else if (url.pathname === "/config") {
       return handleConfig(env, request);
     } else if (url.pathname === "/location") {
-      return handleLocation(env, request);
+      return handleLocation(env, request, corsHeaders(env, request));
     } else if (url.pathname === "/weather") {
-      return handleWeather(env, request);
+      return handleWeather(env, request, corsHeaders(env, request));
     } else if (url.pathname === "/") {
       return handleHealthCheck(env, request);
     } else {
